login error shown and euro decimal fix

+++ OVGPFinalv1/Areas/Identity/Pages/Account/Manage/Index.cshtml.cs
@@ -91,11 +91,11 @@ namespace OVGPFinalv1.Areas.Identity.Pages.Account.Manage
             [PersonalData]
             [DataType(DataType.Currency)]
             [Display(Name = "Betaald bedrag")]
-            public double BetaalBedrag { get; set; }
+            public decimal BetaalBedrag { get; set; }
             [PersonalData]
             [DataType(DataType.Currency)]
             [Display(Name = "Bedrag te voldoen")]
-            public double BedragTeVoldoen { get; set; }
+            public decimal BedragTeVoldoen { get; set; }
         }
 
         public async Task<IActionResult> OnGetAsync()
